+++ crates/assembly-syntax/src/library/module.rs
@@ -8,11 +8,11 @@ use crate::{
     ast::{self, AttributeSet, ConstantValue, Ident, ItemIndex, ProcedureName},
 };
 
-// MODULE INFO
+// MODULE DESCRIPTOR
 // ================================================================================================
 
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ModuleInfo {
+pub struct ModuleDescriptor {
     path: Arc<Path>,
     /// When specified, multiple modules with the same name can be present, so long as they are
     /// disambiguated by version.
@@ -20,8 +20,9 @@ pub struct ModuleInfo {
     items: Vec<ItemInfo>,
 }
 
-impl ModuleInfo {
-    /// Returns a new [`ModuleInfo`] instantiated by library path and optional semantic version.
+impl ModuleDescriptor {
+    /// Returns a new [`ModuleDescriptor`] instantiated by library path and optional semantic
+    /// version.
     ///
     /// The semantic version is optional, as currently the assembler allows assembling artifacts
     /// without providing one.
@@ -130,7 +131,7 @@ impl ModuleInfo {
     }
 }
 
-impl Index<ItemIndex> for ModuleInfo {
+impl Index<ItemIndex> for ModuleDescriptor {
     type Output = ItemInfo;
 
     fn index(&self, index: ItemIndex) -> &Self::Output {