+++ crates/assembly/src/linker/library.rs
@@ -1,6 +1,6 @@
 use alloc::{sync::Arc, vec::Vec};
 
-use miden_assembly_syntax::{Library, library::ModuleInfo};
+use miden_assembly_syntax::{Library, library::ModuleDescriptor};
 use miden_core::mast::MastForest;
 use miden_mast_package::Package;
 use miden_project::Linkage;
@@ -14,7 +14,7 @@ pub struct LinkLibrary {
     /// The MAST of this library
     pub mast: Arc<MastForest>,
     /// Metadata about the modules and symbols available in the linked forest
-    pub module_infos: Vec<ModuleInfo>,
+    pub module_descriptors: Vec<ModuleDescriptor>,
     /// How to link against this library
     pub linkage: Linkage,
 }
@@ -23,9 +23,9 @@ impl LinkLibrary {
     /// Construct a [LinkLibrary] from a [miden_mast_package::Package]
     pub fn from_package(package: Arc<Package>) -> Self {
         let mast = package.mast.mast_forest().clone();
-        let module_infos = package
+        let module_descriptors = package
             .mast
-            .module_infos()
+            .module_descriptors()
             .map(|mut mi| {
                 mi.set_version(package.version.clone());
                 mi
@@ -33,17 +33,17 @@ impl LinkLibrary {
             .collect();
         Self {
             mast,
-            module_infos,
+            module_descriptors,
             linkage: Linkage::Dynamic,
         }
     }
 
     pub(crate) fn from_library(library: &Library) -> Self {
         let mast = library.mast_forest().clone();
-        let module_infos = library.module_infos().collect();
+        let module_descriptors = library.module_descriptors().collect();
         Self {
             mast,
-            module_infos,
+            module_descriptors,
             linkage: Linkage::Dynamic,
         }
     }